Construct filename for still images without using stringstream
For some reason, using a streamstream would cause the following error
when running with `--stream video`: `libav: cannot open  input device`.

This makes absolutely no sense to me, because the `still_save` function
is definitely not being called when `--stream video` is the argument.
Otherwise I suppose this could be a memory issue?